+++ src/projectgoldstarscalculators/QuadraticEquationSolver.java
@@ -7,7 +7,7 @@ import javax.swing.JOptionPane;
 import javax.swing.JTextField;
 public class QuadraticEquationSolver implements ActionListener
 {
-    public static JTextField aField, bField, cField;
+    public static JTextField aField, bField, cField, solutionField;
     
     public QuadraticEquationSolver()
     {
@@ -17,7 +17,7 @@ public class QuadraticEquationSolver implements ActionListener
     private void quadraticEquation()
     {
         ProgramWindow frame = new ProgramWindow("Quadratic Equation Solver");
-        frame.setLayout(new GridLayout(4, 2));
+        frame.setLayout(new GridLayout(5, 2));
         frame.setSize(750 * ProjectGoldStarsCalculatorS.multiplier, 400 * ProjectGoldStarsCalculatorS.multiplier);
         frame.setInstructionsMenuBar("Please enter the quadratic equation in the following form: ax^2 + bx + c = 0");
         frame.add(Components.standardLabel("a:"));
@@ -31,6 +31,9 @@ public class QuadraticEquationSolver implements ActionListener
         frame.add(cField);
         frame.add(new JLabel());
         frame.add(Components.button2("Solve", new SolveQuadraticEquationListener()));
+        frame.add(Components.standardLabel("Solution(s):"));
+        setupSolutionField();
+        frame.add(solutionField);
         frame.makeVisible();
     }
     
@@ -50,6 +53,13 @@ public class QuadraticEquationSolver implements ActionListener
     {
         cField = new JTextField("0");
         cField.setFont(ProjectGoldStarsCalculatorS.bodyText2);
+    }
+    
+    private void setupSolutionField()
+    {
+        solutionField = new JTextField("0");
+        solutionField.setEditable(false);
+        solutionField.setFont(ProjectGoldStarsCalculatorS.bodyText2);
     }
     
     public static class SolveQuadraticEquationListener implements ActionListener
@@ -102,15 +112,11 @@ public class QuadraticEquationSolver implements ActionListener
             x2 = x2 / (2 * aNum);
             if(x1 == x2)    //If there is one unique solution
             {
-                output = "Solving the quadratic equation " + a + "x^2 + " + b + "x + " + c + " = 0...\n"
-                        + "There is 1 possible solution: " + x1;
-                JOptionPane.showMessageDialog(null, output, "Quadratic Equation", JOptionPane.INFORMATION_MESSAGE);
+                solutionField.setText("There is 1 possible solution: " + x1);
             }
             else
             {
-                output = "Solving the quadratic equation " + a + "x^2 + " + b + "x + " + c + " = 0...\n"
-                        + "There are 2 possible solutions: " + x1 + " and " + x2;
-                JOptionPane.showMessageDialog(null, output, "Quadratic Equation", JOptionPane.INFORMATION_MESSAGE);
+                solutionField.setText("There are 2 possible solutions: " + x1 + " and " + x2);
             }
         }
     }